temperature/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.46
urllib3==1.26.13
Werkzeug==2.2.2
pytz
tzlocal
